Changed to make the code compile again. Correctness comes tomorrow ;-)
git-svn-id: https://www.nest-initiative.uni-freiburg.de/nest/branches/10kproject@9230 4fa6ec7b-3718-0410-b4d6-90ca9082ef84

+++ nestkernel/connection_manager.h
@@ -212,7 +212,7 @@ bool ConnectionManager::has_user_prototypes() const
 inline
 int ConnectionManager::get_syn_vec_index(thread tid, index gid, index syn_id) const
 {
-  if (tid >= connections_.size() || gid >= connections_[tid].size() || connections_[tid][gid].size() == 0)
+  if (static_cast<size_t>(tid) >= connections_.size() || gid >= connections_[tid].size() || connections_[tid][gid].size() == 0)
     return -1;
 
   index syn_vec_index = 0;
